Improved parallelisation, more limit options
Numerous updates, most notable of which is the change to the integration function being only for a single particle, and taking the loop outside for better parallelisation. A number of additional input options are available to help kerb runaway particles, such as a cap on how many cells it can pass through. Relaxation has been introduced for the 2nd order scheme, based on iteration number. For the first 6 iterations, force is relaxed until the last 4 iterations have no relaxation, so end result should not be changed, but sensitive particles are brought in more gently. Starting velocity for each iteration is now based on a sliding scale of particle diameter, with small particles based on the cell velocity, and heavy particles based on the previous particle velocity.